eNSP-Automation/requirements.txt
requests>=2.28.0
jinja2>=3.0.0
pywin32>=305
PyQt5>=5.15.0
pythoncom>=0.0.1
logging>=0.5.1.2
jsonschema>=4.17.0
numpy>=1.20.0
python-dotenv>=0.20.0
argparse>=1.4.0
pillow>=9.0.0  # 用于GUI中的图像处理
websocket-client>=1.5.1  # 用于讯飞Spark API的WebSocket通信
netmiko>=4.0.0
paramiko>=2.10.1
cryptography>=39.0.1
pyyaml>=6.0.0
textfsm>=1.1.2
pytest>=7.0.0 